+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/web/views/CreateOfferView.groovy
@@ -3,11 +3,15 @@ package life.qbic.portal.qoffer2.web.views
 import com.vaadin.ui.FormLayout
 import com.vaadin.ui.VerticalLayout
 import life.qbic.datamodel.dtos.business.ProductItem
+import life.qbic.portal.portlet.customers.create.CreateCustomer
+import life.qbic.portal.qoffer2.DependencyManager
+import life.qbic.portal.qoffer2.web.controllers.CreateCustomerController
 import life.qbic.portal.portlet.customers.affiliation.create.CreateAffiliation
 import life.qbic.portal.qoffer2.web.controllers.CreateAffiliationController
 import life.qbic.portal.qoffer2.web.controllers.CreateCustomerController
 import life.qbic.portal.qoffer2.web.controllers.CreateOfferController
 import life.qbic.portal.qoffer2.web.controllers.ListProductsController
+import life.qbic.portal.qoffer2.web.viewmodel.CreateCustomerViewModel
 import life.qbic.portal.qoffer2.web.viewmodel.CreateAffiliationViewModel
 import life.qbic.portal.qoffer2.web.viewmodel.CreateOfferViewModel
 import life.qbic.portal.qoffer2.web.viewmodel.ProductItemViewModel
@@ -33,6 +37,7 @@ class CreateOfferView extends FormLayout{
 
     final private ViewModel sharedViewModel
     final private CreateOfferViewModel view
+
     final private CreateAffiliationViewModel createAffiliationViewModel
 
     final private CreateOfferController controller
@@ -45,15 +50,17 @@ class CreateOfferView extends FormLayout{
     final private SelectItemsView selectItemsView
     final private OfferOverviewView overviewView
 
+    final private CreateCustomerView createCustomerView
     final private CreateAffiliationView createAffiliationView
 
 
-    CreateOfferView(ViewModel sharedViewModel, CreateOfferViewModel createOfferViewModel, CreateOfferController controller, ListProductsController listProductsController, CreateAffiliationView createAffiliationView) {
+    CreateOfferView(ViewModel sharedViewModel, CreateOfferViewModel createOfferViewModel, CreateOfferController controller, ListProductsController listProductsController, CreateCustomerView createCustomerView) {
         super()
         this.sharedViewModel = sharedViewModel
         this.view = createOfferViewModel
         this.controller = controller
         this.listProductsController = listProductsController
+        this.createCustomerView = createCustomerView
         this.createAffiliationView = createAffiliationView
 
         projectInformationView = new ProjectInformationView(view)
@@ -89,6 +96,18 @@ class CreateOfferView extends FormLayout{
         this.customerSelectionView.previous.addClickListener({
             this.removeComponent(customerSelectionView)
             this.addComponent(projectInformationView)
+        })
+        this.customerSelectionView.createCustomerButton.addClickListener({
+            this.removeComponent(customerSelectionView)
+            this.addComponent(createCustomerView)
+        })
+        this.createCustomerView.abortButton.addClickListener({
+            this.removeComponent(createCustomerView)
+            this.addComponent(customerSelectionView)
+        })
+        this.createCustomerView.submitButton.addClickListener({
+            this.removeComponent(createCustomerView)
+            this.addComponent(customerSelectionView)
         })
         this.customerSelectionView.createAffiliationButton.addClickListener({
             this.removeComponent(customerSelectionView)