+++ CHANGELOG.rst
@@ -9,7 +9,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 ----------------------------------------------
 
 **Added**
-* Connect CreateAffilation View and Select Customer View
+* Add the option to create an affiliation while creating an offer
+* Add the option to create a customer while creating an offer
 * Implement SearchCustomer use case
 * Connect CreateAffiliation and CreateCustomer use case
 * Implement GUI for CreateAffiliation use case